pyproject.toml
[tool.poetry]
authors = ["Joachim Dehli", "Lawrence Akka"]
description = "A package for interacting at a low(ish) level with a Loxone miniserver via the Loxone websocket API"
exclude = ["pyloxone_api/lox_token.cfg"]
homepage = "https://github.com/JoDehli/pyloxone-api"
license = "Apache-2.0"
name = "pyloxone_api"
repository = "https://github.com/JoDehli/pyloxone-api"
version = "0.2.4"
# Trove classifiers for PyPI. Licence and Language are set
# automatically
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Home Automation",
]
readme = "README.md"

[tool.poetry.dependencies]
httpx = "^0.23.1"
pycryptodome = "^3.16.0"
python = ">=3.8.1, <4.0"
websockets = "^10.4"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/JoDehli/pyloxone-api/issues"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.10.1"
mypy = "^0.991"
pylint = "^2.15.8"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
pytest-httpx = "^0.21.2"

[tool.poetry.group.test.dependencies]
mypy = "^0.991"

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.3.0"]
